+++ HISTORY.rst
@@ -2,6 +2,20 @@
 History
 =======
 
+0.1.2 (2018-03-12)
+------------------
+
+* The pagination calculation class can now be configured in ``Listing`` and
+  ``Resource`` classes. This enables the use of different pagination defaults
+  in different lists.
+
+
+0.1.1 (2018-03-12)
+------------------
+
+* Untangled Pagination configuration from includeme() function. Pagination
+  (sub-) classes can now be configured via the ``configure()`` method
+
 
 0.1.0 (2018-03-11)
 ------------------